arch-vega: Implement V_MFMA_LD_SCALE_B32
This instruction sets a scale value that is intended to be used for the
*next* MFMA instruction. Currently this only stores the value. The value
is cleared when an MFMA gets the scale value. This is implemented
differently than the spec to avoid massive changes to gem5 to support a
4-dword instruction. Instead we implement the instruction as a VOP3P
instruction and assume the next instruction will be an MFMA instruction.