configs: Fix ruby_fs.py platform after PCI change (#2641)
A recent PR changed they way PCI DMA devices are connected with ruby. As
their DMA port is now connected to the PCI bridge, there is no reason
for saving their DMA port when connecting things. In fact the only DMA
device interfacing with ruby will be the PCI-HOST bridge.


Change-Id: I9e62172e79c94683020ce3194e07929000867e47

Signed-off-by: Giacomo Travaglini <giacomo.travaglini.arm.com>
Co-authored-by: Giacomo Travaglini <giacomo.travaglini.arm.com>